I HATE CYLINDERS I HATE CLYINDERS

+++ ComputerGraphics/Mesh.cpp
@@ -303,7 +303,7 @@ void Mesh::InitialiseCone(float radius, float height, unsigned int segments)
 		const int bottomRightIndex = vertexOffset;
 		const int bottomLeftIndex = vertexOffset + 1 > vertexCount-1 ? 2 : vertexOffset+1;
 
-		const int indexOffset = i * 6;
+		const int indexOffset = (int)i * 6;
 
 		// Top face
 		indices[indexOffset + 0] = topCenterVertexIndex;
